#50 refactor: path 参照元を新 _stores/path/ へ切替
- hook (usePathStore) は新 _stores/path/ (実装/context) を参照
- component の DEFAULT_PATH は _stores/path/constants へ直接参照
- position-store.ts・intent-store.ts の PathStore 型は _types 集約 index 経由に統一
  (旧 _stores/path-store.ts・_contexts/path-store-context.tsx 自体はまだ削除しない)
