+++ src/prototypes/time-control/time-control-03/_components/action-bar/index.tsx
@@ -1,12 +1,12 @@
 import { Button } from '@/components/ui/button'
 
-import { usePathStore } from '../../_contexts/path-store-context'
 import { usePlannedPathStore } from '../../_contexts/planned-path-store-context'
 import { usePositionStore } from '../../_contexts/position-store-context'
 import { useActorStore } from '../../_stores/actor'
 import { useActorSettingsStore } from '../../_stores/actor-settings'
 import { DEFAULT_ACTOR_SETTINGS } from '../../_stores/actor-settings/constants'
 import { useGameClockStore } from '../../_stores/game-clock'
+import { usePathStore } from '../../_stores/path'
 import { ActorId } from '../../types'
 
 type ActionBarProps = {
