+++ src/prototypes/time-control/time-control-03/_components/actor-controller/index.tsx
@@ -5,13 +5,13 @@ import { Button } from '@/components/ui/button'
 import { formatCoord } from '../../../time-control-02/_lib/format-coord'
 import { getTickMs } from '../../../time-control-02/_lib/get-tick-ms'
 import { useIntentStore } from '../../_contexts/intent-store-context'
-import { usePathStore } from '../../_contexts/path-store-context'
 import { usePositionStore } from '../../_contexts/position-store-context'
 import { useActorStore } from '../../_stores/actor'
 import { DEFAULT_ACTOR_INFO } from '../../_stores/actor/constants'
 import { useActorSettingsStore } from '../../_stores/actor-settings'
 import { DEFAULT_ACTOR_SETTINGS } from '../../_stores/actor-settings/constants'
-import { DEFAULT_PATH } from '../../_stores/path-store'
+import { usePathStore } from '../../_stores/path'
+import { DEFAULT_PATH } from '../../_stores/path/constants'
 import { DEFAULT_POSITION } from '../../_stores/position-store'
 import { ActorId } from '../../types'
