+++ src/prototypes/time-control/time-control-03/_lib/build-schedule.ts
@@ -12,7 +12,7 @@ import { ActorId, MovePath, ScheduleRow } from '../types'
  * - 複数 actor 間の実際の実行順序 (auto 進行の実時間ずれ) までは正確にシミュレートしない近似値
  *
  * @param actorIds 対象 actor 一覧
- * @param pathById actor ごとの残り経路 (`_stores/path-store.ts`)
+ * @param pathById actor ごとの残り経路 (`_stores/path/store.ts`)
  * @param actorById actor ごとの移動速度・tick 発生頻度 (`_stores/actor/store.ts`)
  * @param commonGameTimeMs 共通ゲームクロックの現在値 (`_stores/game-clock/store.ts`)
  */
